rename linux -> unix
