+++ src/io/mod.rs
@@ -1,13 +1,13 @@
 #[cfg(target_os = "linux")]
-mod linux;
+mod unix;
 #[cfg(target_os = "linux")]
-pub(crate) use linux::File;
+pub(crate) use unix::File;
 #[cfg(target_os = "linux")]
-pub use linux::IODriver;
+pub use unix::IODriver;
 
 #[cfg(all(target_os = "macos", not(feature = "async-io-rio")))]
 mod linux;
 #[cfg(all(target_os = "macos", not(feature = "async-io-rio")))]
-pub(crate) use linux::File;
+pub(crate) use unix::File;
 #[cfg(all(target_os = "macos", not(feature = "async-io-rio")))]
-pub use linux::IODriver;
+pub use unix::IODriver;
